+++ pleroma_bot/_pin.py
@@ -6,6 +6,7 @@ import requests
 
 from . import logger
 from .i18n import _
+from pleroma_bot._twitter import twitter_api_request
 
 
 def pin_misskey(self, id_post):
@@ -245,8 +246,8 @@ def _get_pinned_tweet_id(self):
         "expansions": "pinned_tweet_id",
         "tweet.fields": "entities",
     }
-    response = requests.get(
-        url, headers=self.header_twitter, params=params, auth=self.auth
+    response = twitter_api_request(
+        'GET', url, headers=self.header_twitter, params=params, auth=self.auth
     )
     if not response.ok:
         response.raise_for_status()